LPS-91856 Move constants and logic from FragmentEntryTypeConstants into FragmentConstants
